+++ src/lib/gadgets/common.ts
@@ -2,7 +2,7 @@ import { Field, FieldConst, FieldVar, VarField } from '../field.js';
 import { Tuple, TupleN } from '../util/types.js';
 import { Snarky } from '../../snarky.js';
 import { MlArray } from '../ml/base.js';
-import { UInt8 } from '../../index.js';
+import { Bool, UInt8 } from '../../index.js';
 import { Provable } from '../provable.js';
 import { chunk } from '../util/arrays.js';
 
@@ -69,8 +69,14 @@ function toVars<T extends Tuple<Field | bigint>>(
   return Tuple.map(fields, toVar);
 }
 
-function assert(stmt: boolean, message?: string): asserts stmt {
-  if (!stmt) {
+/**
+ * Assert that a statement is true. If the statement is false, throws an error with the given message.
+ * Can be used in provable code.
+ */
+function assert(stmt: boolean | Bool, message?: string): asserts stmt {
+  if (stmt instanceof Bool) {
+    stmt.assertTrue(message ?? 'Assertion failed');
+  } else if (!stmt) {
     throw Error(message ?? 'Assertion failed');
   }
 }